+++ src/components/Button/index.tsx
@@ -1,9 +1,16 @@
-import { ButtonProps } from './types';
+import { ButtonProps } from "./types";
 
 function Button(props: ButtonProps) {
-  const { children } = props;
+  const { children, classStyle, ...rest } = props;
 
-  return <button className="bg-red-500">{children}</button>
+  return (
+    <button 
+      className={classStyle ? classStyle : 'bg-blue-600 rounded-md p-1'}
+      {...rest}
+    >
+      {children}
+    </button>
+  );
 }
 
 export default Button;