+++ src/components/Button/types.ts
@@ -2,4 +2,5 @@ import React from "react";
 
 export interface ButtonProps {
   children: string;
+  classStyle?: string;
 }